+++ NEWS.md
@@ -8,6 +8,7 @@
 - Fix segfault printing activities with empty vectors (#253).
 - Fix source behaviour with missing values, which now stop arrival generation
   in the same way as negative values do (#256).
+- Fix `deactivate()`, which now unschedules future arrivals (#249).
 
 ## New features